refactored the WebController file
