+++ src/main/java/hello/EarthquakeQueryService.java
@@ -1,7 +1,7 @@
 package hello;
 
-
-import java.util.Arrays;
+ 
+ import java.util.Arrays;
 
  import org.slf4j.Logger;
  import org.slf4j.LoggerFactory;
