+++ src/main/java/hello/EqSearch.java
@@ -1,4 +1,7 @@
 package hello;
+
+
+
 public class EqSearch {
     private int distance;
     private int minmag;
