+++ src/test/java/hello/EarthquakeSearchTest.java
@@ -21,10 +21,10 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.junit.Before;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.authentication;
-
+import hello.controllers.EarthquakesController;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(WebController.class)
+@WebMvcTest(EarthquakesController.class)
 public class EarthquakeSearchTest {
 
     @Autowired
